Move equipment types to common/rooms/
