+++ indico/modules/rb_new/client/js/containers/FilterBar.jsx
@@ -16,14 +16,14 @@
  */
 import {connect} from 'react-redux';
 import {setFilterParameter, fetchRooms, fetchMapRooms, fetchBlockings} from '../actions';
-import * as selectors from '../selectors';
 import {selectors as userSelectors} from '../common/user';
+import {selectors as roomsSelectors} from '../common/rooms';
 
 
 export default (namespace, componentClass) => {
     const mapStateToProps = state => ({
-        equipmentTypes: selectors.getEquipmentTypes(state),
         ...state[namespace].filters,
+        equipmentTypes: roomsSelectors.getEquipmentTypes(state),
         hasOwnedRooms: userSelectors.hasOwnedRooms(state),
         hasFavoriteRooms: userSelectors.hasFavoriteRooms(state),
         namespace
